Add better NuGetPackageManager tests
